Add example of Float / Int confusion when type signature is present
In this example, Elm has a conflict between Int and Float. However,
the type signature contains an Int and a Float. Based on the
error message below, I can not understand what Elm thinks
is the definition and which Float or Int is incorrect.

The type annotation for `ngon2` does not match its definition.

3| ngon2 : Float -> Int -> Shape
As I infer the type of values flowing through your program, I see a conflict
between these two types:

    Int

    Float